projects/california_housing_regression/requirements.txt
flytekit>=0.30.3
flytekitplugins-pandera
pandas
pandera[strategies]
s3fs
scipy
sklearn
